third_party/blink/web_tests/external/wpt/dom/events/webkit-animation-end-event-expected.txt

This is a testharness.js-based test.
FAIL onanimationend and onwebkitanimationend are not aliases assert_equals: onwebkitanimationend should initially be null expected (object) null but got (undefined) undefined
FAIL dispatchEvent of a webkitAnimationEnd event does trigger a prefixed event handler or listener assert_equals: onwebkitanimationend must be supported and not previously set expected (object) null but got (undefined) undefined
PASS dispatchEvent of a webkitAnimationEnd event does not trigger an unprefixed event handler or listener
FAIL dispatchEvent of an animationend event does not trigger a prefixed event handler or listener assert_equals: onwebkitanimationend must be supported and not previously set expected (object) null but got (undefined) undefined
FAIL onwebkitanimationend event handler should trigger for an animation assert_equals: onwebkitanimationend must be supported and not previously set expected (object) null but got (undefined) undefined
FAIL onwebkitanimationend event handler should not trigger if an unprefixed event handler also exists assert_equals: onwebkitanimationend must be supported and not previously set expected (object) null but got (undefined) undefined
FAIL onwebkitanimationend event handler should not trigger if an unprefixed listener also exists assert_equals: onwebkitanimationend must be supported and not previously set expected (object) null but got (undefined) undefined
FAIL event types for prefixed and unprefixed animationend event handlers should be named appropriately assert_equals: onwebkitanimationend must be supported and not previously set expected (object) null but got (undefined) undefined
PASS webkitAnimationEnd event listener should trigger for an animation
PASS webkitAnimationEnd event listener should not trigger if an unprefixed listener also exists
PASS webkitAnimationEnd event listener should not trigger if an unprefixed event handler also exists
PASS event types for prefixed and unprefixed animationend event listeners should be named appropriately
PASS webkitAnimationEnd event listener is case sensitive
Harness: the test ran to completion.
